Break out VKRFramebuffer/VKRRenderpass from VulkanRenderManager
