+++ Common/Common.vcxproj.filters
@@ -455,6 +455,9 @@
     <ClInclude Include="GPU\MiscTypes.h">
       <Filter>GPU</Filter>
     </ClInclude>
+    <ClInclude Include="GPU\Vulkan\VulkanFramebuffer.h">
+      <Filter>GPU\Vulkan</Filter>
+    </ClInclude>
   </ItemGroup>
   <ItemGroup>
     <ClCompile Include="ABI.cpp" />
@@ -859,6 +862,9 @@
     </ClCompile>
     <ClCompile Include="Render\ManagedTexture.cpp">
       <Filter>Render</Filter>
+    </ClCompile>
+    <ClCompile Include="GPU\Vulkan\VulkanFramebuffer.cpp">
+      <Filter>GPU\Vulkan</Filter>
     </ClCompile>
   </ItemGroup>
   <ItemGroup>
